Create and apply the eventEmitter object This is done using the extendObject utility.
